+++ front/app/components/SignUpIn/SignUpInModal.tsx
@@ -31,19 +31,13 @@ const Container = styled.div``;
 interface Props {
   className?: string;
   onMounted?: () => void;
-  onDeclineInvitation?: () => void;
+  onClosed: () => void;
   onOpened?: (opened: boolean) => void;
   fullScreenModal?: boolean;
 }
 
 const SignUpInModal = memo<Props>(
-  ({
-    className,
-    onMounted,
-    onDeclineInvitation,
-    onOpened,
-    fullScreenModal,
-  }) => {
+  ({ className, onMounted, onClosed, onOpened, fullScreenModal }) => {
     const isMounted = useIsMounted();
     const [metaData, setMetaData] = useState<ISignUpInMetaData | undefined>(
       undefined
@@ -109,15 +103,15 @@ const SignUpInModal = memo<Props>(
         trackEventByName(tracks.signUpFlowExitedAtEmailVerificationStep);
       }
 
-      if (onDeclineInvitation && metaData?.isInvitation) {
-        onDeclineInvitation();
-      }
+      onClosed();
 
       closeSignUpInModal();
     };
 
     const onSignUpInCompleted = () => {
       closeSignUpInModal();
+      onClosed();
+
       const requiresVerification = !!metaData?.verification;
 
       const authUserIsVerified =